Add opacity settings for notes and calendar widget
