+++ androidApp/src/main/AndroidManifest.xml
@@ -30,11 +30,6 @@
         android:theme="@style/Theme.MyBrain"
         tools:ignore="UnusedAttribute">
 
-        <!-- Enables One UI's extended widget handling, including native widget background blur. -->
-        <meta-data
-            android:name="SamsungBasicInteraction"
-            android:value="SEP10" />
-
         <property android:name="android.app.appfunctions.app_metadata"
                   android:resource="@xml/app_metadata" />
 
@@ -90,6 +85,26 @@
 
         <activity
             android:name=".presentation.widget.TasksWidgetConfigurationActivity"
+            android:excludeFromRecents="true"
+            android:exported="true"
+            android:theme="@style/Theme.MyBrain">
+            <intent-filter>
+                <action android:name="android.appwidget.action.APPWIDGET_CONFIGURE" />
+            </intent-filter>
+        </activity>
+
+        <activity
+            android:name=".presentation.widget.NotesWidgetConfigurationActivity"
+            android:excludeFromRecents="true"
+            android:exported="true"
+            android:theme="@style/Theme.MyBrain">
+            <intent-filter>
+                <action android:name="android.appwidget.action.APPWIDGET_CONFIGURE" />
+            </intent-filter>
+        </activity>
+
+        <activity
+            android:name=".presentation.widget.CalendarWidgetConfigurationActivity"
             android:excludeFromRecents="true"
             android:exported="true"
             android:theme="@style/Theme.MyBrain">
