+++ composeApp/src/androidMain/kotlin/com/mhss/app/mybrain/presentation/widget/NotesWidgetConfigurationActivity.kt
@@ -1,0 +1,9 @@
+package com.mhss.app.mybrain.presentation.widget
+
+import com.mhss.app.ui.R
+import com.mhss.app.widget.notes.NotesWidget
+
+class NotesWidgetConfigurationActivity : WidgetConfigurationActivity() {
+    override val widget = NotesWidget()
+    override val titleResource = R.string.notes_widget_settings
+}
