+++ core/ui/src/androidMain/res/values-ar/strings.xml
@@ -7,6 +7,8 @@
     <string name="notes">الملاحظات</string>
     <string name="tasks">المهام</string>
     <string name="tasks_widget_settings">إعدادات أداة المهام</string>
+    <string name="notes_widget_settings">إعدادات أداة الملاحظات</string>
+    <string name="calendar_widget_settings">إعدادات أداة التقويم</string>
     <string name="widget_background_opacity">عتامة الخلفية</string>
     <string name="widget_background_opacity_description">تتيح العتامة المنخفضة ظهور خلفية جهازك من خلالها. وتطبّق أجهزة سامسونج المدعومة أيضًا تمويه أدوات One UI.</string>
     <string name="save">حفظ</string>
