+++ core/ui/src/androidMain/res/values/strings.xml
@@ -7,6 +7,8 @@
     <string name="notes">Notes</string>
     <string name="tasks">Tasks</string>
     <string name="tasks_widget_settings">Tasks widget settings</string>
+    <string name="notes_widget_settings">Notes widget settings</string>
+    <string name="calendar_widget_settings">Calendar widget settings</string>
     <string name="widget_background_opacity">Background opacity</string>
     <string name="widget_background_opacity_description">Lower opacity lets your wallpaper show through. Supported Samsung devices also apply One UI widget blur.</string>
     <string name="save">Save</string>
