+++ core/ui/src/androidMain/res/xml/notes_widget_info.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:resizeMode="horizontal|vertical"
+    android:configure="com.mhss.app.mybrain.presentation.widget.NotesWidgetConfigurationActivity"
     android:initialLayout="@layout/widget_initial"
     android:previewImage="@drawable/notes_widget_preview_img"
     android:minHeight="110dp"
@@ -8,5 +10,9 @@
     android:minResizeWidth="110dp"
     android:minWidth="110dp"
     android:widgetCategory="home_screen"
+    android:widgetFeatures="reconfigurable|configuration_optional"
+    app:previewLayoutLarge="@layout/notes_widget_preview"
+    app:previewSize="large"
+    app:widgetSize="small|wideSmall|medium|large|extraLarge|extraLargeLong"
+    app:widgetStyle="colorful"
 />
-
