+++ widget/src/main/java/com/mhss/app/widget/calendar/CalendarWidget.kt
@@ -7,11 +7,13 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.core.content.ContextCompat
+import androidx.datastore.preferences.core.Preferences
 import androidx.glance.GlanceId
 import androidx.glance.GlanceTheme
 import androidx.glance.appwidget.GlanceAppWidget
 import androidx.glance.appwidget.GlanceAppWidgetReceiver
 import androidx.glance.appwidget.provideContent
+import androidx.glance.currentState
 import androidx.glance.material3.ColorProviders
 import com.mhss.app.datetime.DateTimeFormatter
 import com.mhss.app.datetime.LocalDateTimeFormatter
@@ -23,6 +25,7 @@ import com.mhss.app.preferences.domain.model.stringSetPreferencesKey
 import com.mhss.app.preferences.domain.use_case.GetPreferenceUseCase
 import com.mhss.app.ui.ThemeSettings
 import com.mhss.app.ui.toIntList
+import com.mhss.app.widget.WidgetSettings
 import com.mhss.app.widget.WidgetTheme
 import com.mhss.app.widget.widgetDarkColorScheme
 import com.mhss.app.widget.widgetLightColorScheme
@@ -45,6 +48,8 @@ class CalendarWidget : GlanceAppWidget(), KoinComponent {
         val events = getAllEvents(includedCalendars.toIntList(), fromWidget = true)
 
         provideContent {
+            val widgetPreferences = currentState<Preferences>()
+            val backgroundOpacity = WidgetSettings.backgroundOpacity(widgetPreferences)
             val useMaterialYou by getSettings(
                 booleanPreferencesKey(PrefsConstants.SETTINGS_MATERIAL_YOU),
                 false
@@ -80,7 +85,8 @@ class CalendarWidget : GlanceAppWidget(), KoinComponent {
                 ) {
                     CalendarHomeScreenWidget(
                         events.eventDays,
-                        hasPermission
+                        hasPermission,
+                        backgroundOpacity
                     )
                 }
             }
