+++ widget/src/main/java/com/mhss/app/widget/notes/NotesWidget.kt
@@ -5,11 +5,13 @@ import android.content.res.Configuration
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
+import androidx.datastore.preferences.core.Preferences
 import androidx.glance.GlanceId
 import androidx.glance.GlanceTheme
 import androidx.glance.appwidget.GlanceAppWidget
 import androidx.glance.appwidget.GlanceAppWidgetReceiver
 import androidx.glance.appwidget.provideContent
+import androidx.glance.currentState
 import androidx.glance.material3.ColorProviders
 import com.mhss.app.domain.use_case.GetAllNotesUseCase
 import com.mhss.app.preferences.PrefsConstants
@@ -21,6 +23,7 @@ import com.mhss.app.preferences.domain.model.toInt
 import com.mhss.app.preferences.domain.model.toOrder
 import com.mhss.app.preferences.domain.use_case.GetPreferenceUseCase
 import com.mhss.app.ui.ThemeSettings
+import com.mhss.app.widget.WidgetSettings
 import com.mhss.app.widget.WidgetTheme
 import com.mhss.app.widget.widgetDarkColorScheme
 import com.mhss.app.widget.widgetLightColorScheme
@@ -35,6 +38,8 @@ class NotesWidget : GlanceAppWidget(), KoinComponent {
     override suspend fun provideGlance(context: Context, id: GlanceId) {
 
         provideContent {
+            val widgetPreferences = currentState<Preferences>()
+            val backgroundOpacity = WidgetSettings.backgroundOpacity(widgetPreferences)
             val order by getSettings(
                 intPreferencesKey(PrefsConstants.NOTES_ORDER_KEY),
                 Order.DateModified(OrderType.ASC).toInt()
@@ -72,7 +77,8 @@ class NotesWidget : GlanceAppWidget(), KoinComponent {
                 else ColorProviders(widgetLightColorScheme)
             ) {
                 NotesHomeScreenWidget(
-                    limitedNotes
+                    notes = limitedNotes,
+                    backgroundOpacity = backgroundOpacity
                 )
             }
         }
